feat: implement comprehensive integer-only financial architecture
Implements complete integer-only financial data architecture eliminating all
floating-point arithmetic throughout the system:

Database Schema:
- Convert all tax rates from decimal to unsigned integers (basis points: 18% = 1800)
- Convert monetary fields to unsigned big integers (cents: $100.00 = 10000)
- Convert exchange rates to unsigned big integers (micro-units: 1.234567 = 1234567)
- Optimize field types (char vs string) for better performance

Models & Business Logic:
- Update InvoiceItem: tax calculations using pure integer arithmetic
- Update TaxTemplate: store/retrieve basis points with formatted display
- Update Invoice: exchange rate handling with custom cast
- Create ExchangeRateCast for seamless decimal <-> micro-units conversion
- Remove all decimal/float casts and calculations from business logic

User Interface:
- InvoiceWizard: convert between user input (%) and storage (basis points)
- Tax rate formatters: 1800 basis points → "18.00%" display
- Money formatters: integer cents → currency display strings
- Exchange rate formatters: micro-units → decimal strings

Services & Components:
- EstimateToInvoiceConverter: include currency/exchange_rate fields
- CreateNewUser: set default currency for new teams
- InvoiceCalculator: pure integer arithmetic throughout

Testing:
- Update all test expectations to use basis points (1800) vs percentages (18)
- Comprehensive edge case testing for integer calculations
- Factory updates to generate values in correct scales

Key Benefits:
- Eliminates floating-point precision errors
- Ensures consistent financial calculations
- Maintains backward compatibility through formatting layer
- Improves database performance with optimized field types

+++ database/migrations/2025_07_05_200317_create_invoice_items_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('invoice_id')->constrained()->onDelete('cascade');
             $table->text('description');
-            $table->integer('quantity');
-            $table->integer('unit_price');
-            $table->decimal('tax_rate', 5, 2)->nullable();
+            $table->unsignedInteger('quantity');
+            $table->unsignedBigInteger('unit_price'); // Price in cents
+            $table->unsignedInteger('tax_rate')->nullable(); // Tax rate in basis points (18% = 1800)
             $table->timestamps();
         });
     }
